+++ packages/main/vite.config.js
@@ -1,7 +1,7 @@
 import {join} from 'node:path';
 import {node} from '../../.electron-vendors.cache.json';
-import importCommonPlugin from '../../scripts/importCommonPlugin';
-import {injectAppVersion} from '../../version/inject-app-version-plugin.mjs';
+import importCommonPlugin from '../../plugins/import-common-plugin';
+import {injectAppVersion} from '../../plugins/inject-app-version-plugin.mjs';
 
 const PACKAGE_ROOT = __dirname;
 const PROJECT_ROOT = join(PACKAGE_ROOT, '../..');
