+++ packages/renderer/vite.config.js
@@ -4,7 +4,7 @@ import {chrome} from '../../.electron-vendors.cache.json';
 import react from '@vitejs/plugin-react';
 import {renderer} from 'unplugin-auto-expose';
 import {join} from 'node:path';
-import {injectAppVersion} from '../../version/inject-app-version-plugin.mjs';
+import {injectAppVersion} from '../../plugins/inject-app-version-plugin.mjs';
 
 const PACKAGE_ROOT = __dirname;
 const PROJECT_ROOT = join(PACKAGE_ROOT, '../..');
